+++ adapters/Mintegral/MintegralAdapter/Bidding/GADMAdapterMintegralRTBRewardedAdLoader.h
@@ -15,11 +15,11 @@
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-
-@interface GADMAdapterMintegralRTBRewardedAdLoader : NSObject<GADMediationRewardedAd>
+@interface GADMAdapterMintegralRTBRewardedAdLoader : NSObject <GADMediationRewardedAd>
 
 /// Asks the receiver to load the ad configuration.
-- (void)loadRTBRewardedAdForAdConfiguration:(nonnull GADMediationRewardedAdConfiguration *)adConfiguration
-                       completionHandler:(nonnull GADMediationRewardedLoadCompletionHandler)completionHandler;
+- (void)loadRTBRewardedAdForAdConfiguration:
+            (nonnull GADMediationRewardedAdConfiguration *)adConfiguration
+                          completionHandler:
+                              (nonnull GADMediationRewardedLoadCompletionHandler)completionHandler;
 @end
-