+++ adapters/Mintegral/MintegralAdapter/GADMediationAdapterMintegral.h
@@ -16,16 +16,15 @@
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
 typedef NS_ENUM(NSInteger, GADMintegralErrorCode) {
-    /// Missing server parameters.
-    GADMintegralErrorInvalidServerParameters   = 101,
-    /// The ad request was successful, but no ad was returned.
-    GADMintegralErrorAdNotAvailable = 102,
-    /// The Mintegral SDK failed to show an ad.
-    GADMintegralErrorAdFailedToShow  = 103,
-    /// Invalid banner size for Mintegral ad.
-    GADMintegtalErrorBannerSizeInValid = 104
+  /// Missing server parameters.
+  GADMintegralErrorInvalidServerParameters = 101,
+  /// The ad request was successful, but no ad was returned.
+  GADMintegralErrorAdNotAvailable = 102,
+  /// The Mintegral SDK failed to show an ad.
+  GADMintegralErrorAdFailedToShow = 103,
+  /// Invalid banner size for Mintegral ad.
+  GADMintegtalErrorBannerSizeInValid = 104
 };
 @interface GADMediationAdapterMintegral : NSObject <GADRTBAdapter>
 
 @end
-