+++ adapters/Mintegral/MintegralAdapter/Waterfall/GADMAdapterMintegralInterstitialAdLoader.h
@@ -15,9 +15,11 @@
 #import <Foundation/Foundation.h>
 #import <GoogleMobileAds/GoogleMobileAds.h>
 
-@interface GADMAdapterMintegralNativeAdLoader : NSObject<GADMediationNativeAd>
+@interface GADMAdapterMintegralInterstitialAdLoader : NSObject <GADMediationInterstitialAd>
 
 /// Asks the receiver to render the ad configuration.
-- (void)loadNativeAdForAdConfiguration:(nonnull GADMediationNativeAdConfiguration *)adConfiguration completionHandler:(nonnull GADMediationNativeLoadCompletionHandler)completionHandler;
+- (void)loadInterstitialAdForAdConfiguration:
+            (nonnull GADMediationInterstitialAdConfiguration *)adConfiguration
+                           completionHandler:(nonnull GADMediationInterstitialLoadCompletionHandler)
+                                                 completionHandler;
 @end
-